+++ src/api/domain/searcher/feedly/search_response.go
